Move tool to release command group
