+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/__init__.py
@@ -11,9 +11,8 @@ from .docs import docs
 from .env import env
 from .meta import meta
 from .release import release
-from .release_stats import release_stats
 from .run import run
 from .test import test
 from .validate import validate
 
-ALL_COMMANDS = (agent, ci, clean, config, create, dep, docs, env, meta, release, release_stats, run, test, validate)
+ALL_COMMANDS = (agent, ci, clean, config, create, dep, docs, env, meta, release, run, test, validate)
