+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/__init__.py
@@ -8,11 +8,12 @@ from .build import build
 from .changelog import changelog
 from .make import make
 from .show import show
+from .stats import stats
 from .tag import tag
 from .trello import trello
 from .upload import upload
 
-ALL_COMMANDS = [build, changelog, make, show, tag, trello, upload]
+ALL_COMMANDS = [build, changelog, make, show, stats, tag, trello, upload]
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='Manage the release of checks')
