+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/stats/__init__.py
@@ -3,16 +3,16 @@
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import click
 
-from ..console import CONTEXT_SETTINGS
+from ...console import CONTEXT_SETTINGS
 from .csv_report import csv_report
 
 ALL_COMMANDS = (csv_report,)
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='A collection of tasks to generate reports about releases')
-def release_stats():
+def stats():
     pass
 
 
 for command in ALL_COMMANDS:
-    release_stats.add_command(command)
+    stats.add_command(command)
